Cache keywords to speed up chatting with built-in AI
1. Cache keywords when the user is watching videos
   Move updateContent.js to src/.
   Use craco.
   Extract generateKeywordMap function and use in updateContent.js
2. Generate sessions on demand

TODO: summarizer disabled.
