+++ src/components/ChatComponents/ChatSearch.jsx
@@ -11,9 +11,10 @@ import ArrowUpIcon from '../ArrowUpIcon';
 import { getPromptResult, getStreamingPromptResult } from '../../lib/getPromptResult';
 import { LanguageContext } from '../../context/LanguageContext';
 import { getGeminiResponse } from '../../lib/getGeminiResponse';
+import { generatePromptSession } from '../../lib/generateKeywordMap';
 import { SearchContext } from '../../context/SearchContext';
 
-export default function ChatSearch({ messages, setMessages, timestampedSubtitles, masterPromptSession, promptSessionArray, subTitlesLoading, model, setModel }) {
+export default function ChatSearch({ messages, setMessages, timestampedSubtitles, masterPromptSession, promptSessionArray, setPromptSessionArray, subtitleChunkArray, subTitlesLoading, model, setModel }) {
     const [disabled, setDisabled] = useState(false);
     const [searchQuery, setSearchQuery] = useState('');
     const { convertOutputText, outputLanguage } = useContext(LanguageContext);
@@ -39,6 +40,8 @@ export default function ChatSearch({ messages, setMessages, timestampedSubtitles
             setSearchQuery("");
             if (model == 'chrome-built-in') {
                 try {
+                    console.log("masterPromptSession", masterPromptSession);
+                    console.time("run-masterPrompt");
                     const sessionIdString = await masterPromptSession.prompt(searchQuery);
                     console.log("sessionIdString", sessionIdString);
 
@@ -48,17 +51,27 @@ export default function ChatSearch({ messages, setMessages, timestampedSubtitles
                     console.log("promptSessionId", promptSessionId);
 
                     // Validate and adjust `promptSessionId`
-                    if (Number.isNaN(promptSessionId) || promptSessionId >= promptSessionArray.length) {
+                    if (Number.isNaN(promptSessionId) || promptSessionId >= subtitleChunkArray.length) {
                         promptSessionId = 0;
                     }
+                    console.timeEnd("run-masterPrompt");
+                    console.time("get-generate-session");
+                    let session = promptSessionArray[promptSessionId];
+                    if (!session) {
+                        session = await generatePromptSession(subtitleChunkArray[promptSessionId]);
+                        setPromptSessionArray(prev => ({ ...prev, [promptSessionId]: session }));
+                    }
+                    console.timeEnd("get-generate-session");
+                    console.time("getStreamingPromptResult");
 
                     // Call `getStreamingPromptResult` with a valid `promptSessionId`
                     const UpdatedQuery = searchQuery + " give me the response in the described format, give integer time after each paragraph in curly braces for example {45} etc. Never include time range in curly braces. Give answer in headings, subheadings and bullet points"
                     await getStreamingPromptResult(
-                        promptSessionArray[promptSessionId],
+                        session,
                         UpdatedQuery,
                         handleStreamResponse
                     );
+                    console.timeEnd("getStreamingPromptResult");
 
 
                 } catch (error) {
